Rename points.tsx.
